Moved ChunkData and ChunkInfo structs into chunk module
